update electricityMap Interface and Comments
