--- src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/InternalsVisibleTo.cs
@@ -1,4 +1,0 @@
-using System.Runtime.CompilerServices;
-
-// TODO:Watttime -> electricityMap
-[assembly:InternalsVisibleTo("CarbonAware.Tools.WattTimeClient.Tests")]
