+++ src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/Configuration/electricityMapClientConfiguration.cs
