+++ src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/Constants/AuthenticationHeaderTypes.cs
