+++ src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/Constants/Paths.cs
@@ -4,7 +4,6 @@ internal class Paths
 {
     public const string Data = "data";
     public const string Forecast = "forecast";
-    // public const string ZoneFromLocation = "ba-from-loc";
     public const string Login = "login";
     public const string Historical = "past";
 }
