+++ src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/Constants/QueryStrings.cs
