+++ src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/Model/Forecast.cs
