+++ src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/Model/GridEmissionDataPoint.cs
@@ -3,13 +3,13 @@
 namespace CarbonAware.Tools.electricityMapClient.Model;
 
 /// <summary>
-/// An object describing the emissions for a given time period and balancing authority.
+/// An object describing the emissions for a given countryCode
 /// </summary>
 [Serializable]
 public record GridEmissionDataPoint
 {
     /// <summary>
-    /// Balancing authority abbreviation
+    /// countryCode
     /// </summary>
     [JsonPropertyName("countryCode")]
     public string countryCodeAbbreviation { get; set; } = string.Empty;
@@ -21,7 +21,7 @@ public record GridEmissionDataPoint
     public string? Status { get; set; }
 
     /// <summary>
-    /// TODO: need deserialize or access to a property in a nested object(data) e.g. make another class
+    /// TODO: need deserialize or access to a property in a nested object(data)
     /// </summary>
     [JsonPropertyName("data")]
     public string[] data { get; set; }
