+++ src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/electricityMapClientException.cs
